move graph building back into core lib
