+++ rust/routee-compass-core/src/model/graph/mod.rs
@@ -2,5 +2,6 @@ pub mod edge_id;
 pub mod edge_loader;
 pub mod graph;
 pub mod graph_error;
+pub mod graph_loader;
 pub mod vertex_id;
 pub mod vertex_loader;
